Main header and self photo
I finished with the main header that introduces myself and what my role is. I then added an image of myself with a pink "blob" behind the image for some depth and color. I used absolute positioning for the pink blob and relative for the actual picture to keep it in place when the screen is sized up or down.
The intro has a display of block so that the h1 and h2 stack on top of each other.